pyproject.toml
[project]
name = "roadmaps-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "fastapi[standard]>=0.124.4",
    "langchain>=1.0.8",
    "langchain-google-genai>=3.1.0",
    "langgraph>=1.0.3",
    "prisma>=0.15.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
    "ruff>=0.14.6",
]
[tool.ruff]
line-length = 100

[tool.ruff.format]
quote-style = "single"
indent-style = "tab"
docstring-code-format = true
